@@ -1,7 +1,7 @@
-'use strict';
+
 
 const path = require('path');
 
 module.exports = {
-    publicFolder: path.join(__dirname, 'dist'),
+    publicFolder: path.join(__dirname, 'build'),
 };
