@@ -3,7 +3,8 @@ import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
 
 import { createStore } from 'redux';
-import { mount } from 'enzyme/build';
+
+import { render, screen, fireEvent } from '@testing-library/react';
 
 import Reporting from '../Reporting';
 import { REPORTING_SELECT_ID } from '../../../testIds';
@@ -16,8 +17,8 @@ const mockStore = {
 };
 const mockReducer = state => state;
 
-test('changing projects renders only toggles from that project', () => {
-    const wrapper = mount(
+test('changing projects renders only toggles from that project', async () => {
+    render(
         <HashRouter>
             <Provider store={createStore(mockReducer, mockStore)}>
                 <Reporting projects={testProjects} features={testFeatures} fetchFeatureToggles={() => {}} />
@@ -25,14 +26,9 @@ test('changing projects renders only toggles from that project', () => {
         </HashRouter>
     );
 
-    const select = wrapper.find(`input[data-test="${REPORTING_SELECT_ID}"][value="default"]`).first();
-
-    let list = wrapper.find('tr');
+    const table = await screen.findByRole("table");
     /* Length of projects belonging to project (3) + header row (1) */
-    expect(list.length).toBe(4);
-
-    select.simulate('change', { target: { value: 'myProject' } });
-    list = wrapper.find('tr');
-
-    expect(list.length).toBe(3);
+    expect(table.rows).toHaveLength(4);
+    fireEvent.change(await screen.findByTestId(REPORTING_SELECT_ID), { target: { value: 'myProject'}});
+    expect(table.rows).toHaveLength(3);
 });
