@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
 import { loadInitialData } from '../../../store/loader';
 
-const mapStateToProps = state => ({ uiConfig: state.uiConfig.toJS() });
-
 import Header from './Header';
+
+const mapStateToProps = state => ({ uiConfig: state.uiConfig.toJS() });
 
 export default connect(mapStateToProps, {
     init: loadInitialData,
