@@ -1,24 +1,24 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import {
+    Avatar,
+    Button,
     Card,
     CardHeader,
-    Avatar,
-    List,
-    ListItem,
-    ListItemSecondaryAction,
-    ListItemText,
-    ListItemAvatar,
-    Select,
-    MenuItem,
-    Icon,
-    IconButton,
     Dialog,
     DialogActions,
-    DialogTitle,
-    DialogContentText,
     DialogContent,
-    Button,
+    DialogContentText,
+    DialogTitle,
+    Icon,
+    IconButton,
+    List,
+    ListItem,
+    ListItemAvatar,
+    ListItemSecondaryAction,
+    ListItemText,
+    MenuItem,
+    Select,
 } from '@material-ui/core';
 
 import AddUserComponent from './access-add-user';
@@ -38,6 +38,7 @@ function AccessComponent({ projectId, project }) {
 
     useEffect(() => {
         fetchAccess();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [projectId]);
 
     if (!project) {
