@@ -49,8 +49,8 @@ export default class ShowUserComponent extends React.Component {
         const email = this.props.profile ? this.props.profile.email : '';
         const locale = this.getLocale();
         let foundLocale = this.possibleLocales.find(l => l.value === locale);
-        const imageUrl = email ? this.props.profile.imageUrl : 'public/unknown-user.png';
-        const imageLocale = foundLocale ? `public/${foundLocale.image}.png` : `public/unknown-locale.png`;
+        const imageUrl = email ? this.props.profile.imageUrl : 'unknown-user.png';
+        const imageLocale = foundLocale ? `${foundLocale.image}.png` : `unknown-locale.png`;
         return (
             <div className={styles.showUserSettings}>
                 <DropdownMenu
@@ -60,7 +60,7 @@ export default class ShowUserComponent extends React.Component {
                         this.possibleLocales.map(i => (
                             <MenuItem key={i.value} onClick={() => this.setLocale(i)}>
                                 <div className={styles.showLocale}>
-                                    <img src={`public/${i.image}.png`} title={i.value} alt={i.value} />
+                                    <img src={`${i.image}.png`} title={i.value} alt={i.value} />
                                     <Typography variant="p">{i.value}</Typography>
                                 </div>
                             </MenuItem>
