@@ -1,7 +1,7 @@
 /* eslint-disable no-alert */
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Icon, Table, TableBody, TableCell, TableHead, TableRow } from '@material-ui/core';
+import { Button, Icon, IconButton, Table, TableBody, TableCell, TableHead, TableRow } from '@material-ui/core';
 import { formatFullDateTimeWithLocale } from '../../../../component/common/util';
 import AddUser from '../add-user-component';
 import ChangePassword from '../change-password-component';
@@ -65,6 +65,7 @@ function UsersList({
 
     useEffect(() => {
         fetchUsers();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     return (
@@ -92,22 +93,22 @@ function UsersList({
                                 condition={hasPermission('ADMIN')}
                                 show={
                                     <TableCell>
-                                        <a href="" title="Edit" onClick={openUpdateDialog(item)}>
+                                        <IconButton aria-label="Edit" title="Edit" onClick={openUpdateDialog(item)}>
                                             <Icon>edit</Icon>
-                                        </a>
-                                        <a href="" title="Change password" onClick={openPwDialog(item)}>
+                                        </IconButton>
+                                        <IconButton aria-label="Change password" title="Change password" onClick={openPwDialog(item)}>
                                             <Icon>lock</Icon>
-                                        </a>
-                                        <a href="" title="Remove user" onClick={openDelDialog(item)}>
+                                        </IconButton>
+                                        <IconButton aria-label="Remove user" title="Remove user" onClick={openDelDialog(item)}>
                                             <Icon>delete</Icon>
-                                        </a>
+                                        </IconButton>
                                     </TableCell>
                                 }
                                 elseShow={
                                     <TableCell>
-                                        <a href="" title="Change password" onClick={openPwDialog(item)}>
+                                        <IconButton aria-label="Change password" title="Change password" onClick={openPwDialog(item)}>
                                             <Icon>lock</Icon>
-                                        </a>
+                                        </IconButton>
                                     </TableCell>
                                 }
                             />
