@@ -1,5 +1,6 @@
+import '@testing-library/jest-dom'
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-configure({ adapter: new Adapter() });
 process.env.TZ = 'UTC';
+configure({ adapter: new Adapter() });
