@@ -1,5 +1,5 @@
 import api from './api';
-import { dispatchAndThrow } from '../util';
+import { dispatchError } from '../util';
 
 export const RECEIVE_FEATURE_TYPES = 'RECEIVE_FEATURE_TYPES';
 export const ERROR_RECEIVE_FEATURE_TYPES = 'ERROR_RECEIVE_FEATURE_TYPES';
@@ -11,5 +11,5 @@ export function fetchFeatureTypes() {
         api
             .fetchAll()
             .then(json => dispatch(receiveFeatureTypes(json)))
-            .catch(dispatchAndThrow(dispatch, ERROR_RECEIVE_FEATURE_TYPES));
+            .catch(dispatchError(dispatch, ERROR_RECEIVE_FEATURE_TYPES));
 }
