@@ -1,6 +1,6 @@
 import api from './api';
 import applicationApi from '../application/api';
-import { dispatchAndThrow } from '../util';
+import { dispatchError } from '../util';
 
 export const ADD_STRATEGY = 'ADD_STRATEGY';
 export const UPDATE_STRATEGY = 'UPDATE_STRATEGY';
@@ -44,7 +44,7 @@ export function fetchStrategies() {
         return api
             .fetchAll()
             .then(json => dispatch(receiveStrategies(json)))
-            .catch(dispatchAndThrow(dispatch, ERROR_RECEIVE_STRATEGIES));
+            .catch(dispatchError(dispatch, ERROR_RECEIVE_STRATEGIES));
     };
 }
 
@@ -55,7 +55,7 @@ export function createStrategy(strategy) {
         return api
             .create(strategy)
             .then(() => dispatch(addStrategy(strategy)))
-            .catch(dispatchAndThrow(dispatch, ERROR_CREATING_STRATEGY));
+            .catch(dispatchError(dispatch, ERROR_CREATING_STRATEGY));
     };
 }
 
@@ -66,7 +66,7 @@ export function updateStrategy(strategy) {
         return api
             .update(strategy)
             .then(() => dispatch(updatedStrategy(strategy)))
-            .catch(dispatchAndThrow(dispatch, ERROR_UPDATING_STRATEGY));
+            .catch(dispatchError(dispatch, ERROR_UPDATING_STRATEGY));
     };
 }
 
@@ -75,7 +75,7 @@ export function removeStrategy(strategy) {
         api
             .remove(strategy)
             .then(() => dispatch(createRemoveStrategy(strategy)))
-            .catch(dispatchAndThrow(dispatch, ERROR_REMOVING_STRATEGY));
+            .catch(dispatchError(dispatch, ERROR_REMOVING_STRATEGY));
 }
 
 export function getApplicationsWithStrategy(strategyName) {
@@ -87,7 +87,7 @@ export function deprecateStrategy(strategy) {
         dispatch(startDeprecate());
         api.deprecate(strategy)
             .then(() => dispatch(deprecateStrategyEvent(strategy)))
-            .catch(dispatchAndThrow(dispatch, ERROR_DEPRECATING_STRATEGY));
+            .catch(dispatchError(dispatch, ERROR_DEPRECATING_STRATEGY));
     };
 }
 
@@ -96,6 +96,6 @@ export function reactivateStrategy(strategy) {
         dispatch(startReactivate());
         api.reactivate(strategy)
             .then(() => dispatch(reactivateStrategyEvent(strategy)))
-            .catch(dispatchAndThrow(dispatch, ERROR_REACTIVATING_STRATEGY));
+            .catch(dispatchError(dispatch, ERROR_REACTIVATING_STRATEGY));
     };
 }
